% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_score.R
\name{dummy_score}
\alias{dummy_score}
\title{Dummy Extension Scoring}
\usage{
dummy_score(
  data,
  fa = NULL,
  loadings = fa$loadings,
  method = c("Rubin", "Bartlett", "Harman", "Anderson"),
  parallel = F,
  core.variables = NULL,
  scoring.variables = NULL,
  corr.matrix = NULL,
  ...
)
}
\arguments{
\item{data}{A matrix or dataframe containing the data to be analyzed}

\item{fa}{(optional) An object of class 'fa' from the 'fa' function in the 'psych' package, used to define the factors to be scored}

\item{loadings}{(optional) Alternatively, a matrix of factor loadings}

\item{method}{Which Dummy-Extension method to use. Options are 'Rubin','Bartlett','Harman', and 'Anderson.' Rubin is similar to the Thurstone method in traditional factor analysis, and Bartlett and Harman are precisely the same. 'Anderson' is short for the Anderson/Rubin method that produces uncorrelated factor scores.}

\item{parallel}{A logical indicating whether to use multiple cores for scoring. This will only offer speed improvement for very large datasets, and will probably be slower for most datasets.}

\item{core.variables}{(optional) Can be either a vector of variable names or index numbers. Defines the 'core' variables from which the factor extension is done. If 'fa' and 'loadings' and 'corr.matrix' are all null, it will use all the variables in 'data'}

\item{scoring.variables}{(optional) Either a vector of variable names of index numbers. Defines the variables used in the scoring step, which can be 'core' variables or other variables, or any combination thereof. Defaults to all the variables in 'data'}

\item{corr.matrix}{(optional) A correlation matrix of the 'core' variables, used in the factor extension step}

\item{...}{(optional) Arguments passed to the psych::fa function if no factor loadings are provided. See psych::fa for more details.}
}
\value{
A list with elements 'scores,' 'original.loadings,' and 'extension.loadings.' The factor scores for every observation are contained in 'scores.' The loadings of the core variables are in 'original.loadings' and the category loadings of the scoring variables are in 'extension.loadings'
}
\description{
Scores a matrix or dataframe on latent factors, using Dummy Extension Scoring (Keenan, 2023). The factors to be scored can either be provided in the form of a psych::fa output, or a matrix of loadings. Alternatively, if only a data matrix is provided, the factor analysis will be performed within the dummy_score function, with parameter passed by '...' The user can select 'core.variables' to be factored, and also 'scoring.variables' to be used in the scoring step. The 'scoring.variables' can include factored variables, variables not factored, or any combination thereof.
}
\examples{
\dontrun{big.five.scores <- dummy_score(data = bigfive, method = 'Bartlett')$scores}
}
